+++ src/app/components/tables/data-table/cell/terra-data-table-header-cell.interface.ts
@@ -6,8 +6,11 @@ import { TerraTextAlignEnum } from './terra-text-align.enum';
 export interface TerraDataTableHeaderCellInterface
 {
     caption:string;
-    identifier:string;
     width:number;
+    /**
+     * this parameter should state the name of the corresponding data base table cell
+     */
+    sortBy?:string;
     isHidden?:boolean;
     tooltipText?:string;
     tooltipPlacement?:string;