+++ src/app/components/tables/data-table/cell/terra-ref-type.interface.ts
@@ -2,12 +2,14 @@ export enum TerraRefTypeEnum
 {
     email = 'mailto',
     phone = 'tel',
+    url = 'url',
     function = 'function'
 }
 
 export interface TerraRefTypeInterface
 {
     type:TerraRefTypeEnum;
-    value:string;
-    func?:() => void;
+    value:string | number | Function;
+    caption?:any;
+    target?:string;
 }